requirements.txt
fastapi
SQLAlchemy
pydantic
python-dotenv
azure-ai-inference
sqlalchemy-utils
pgvector
nltk
PyPDF2
python-docx
Markdown
beautifulsoup4
pytesseract
pillow
pymupdf
psycopg2-binary
python-multipart
uvicorn
openai